move warning logic to UpgradeNotificationHelper
